refactor: incident details sidebar refactored to extract responders panel & incident details panel components
