+++ src/components/IncidentDetails/DefinitionOfDone/EditEvidenceDefinitionOfDoneScript.tsx
@@ -34,6 +34,7 @@ export default function EditEvidenceDefinitionOfDoneScript({
       size="full"
       actions={[
         <button
+          key="update"
           className="px-4 py-2 btn-primary"
           type="button"
           onClick={async () => {
